Hook up vDOM component rendering higher up
