+++ bokehjs/src/lib/core/dom.ts
@@ -2,9 +2,14 @@ import {isBoolean, isNumber, isString, isArray, isPlainObject} from "./util/type
 import {entries} from "./util/object"
 import {BBox} from "./util/bbox"
 import type {Size, Box, Extents, PlainObject} from "./types"
-import type {CSSStyles, CSSStyleSheetDecl} from "./css"
-import {compose_stylesheet, apply_styles} from "./css"
+import type {CSSStyles} from "./css"
+import {apply_styles} from "./css"
 import {logger} from "./logging"
+
+// {{{ TODO remove this
+export type {StyleSheet, StyleSheetLike} from "./stylesheets"
+export {InlineStyleSheet, ImportedStyleSheet, GlobalInlineStyleSheet, GlobalImportedStyleSheet} from "./stylesheets"
+// }}}
 
 export type Optional<T> = {[P in keyof T]?: T[P] | null | undefined}
 
@@ -686,119 +691,6 @@ export enum MouseButton {
   Right = Secondary,
   Middle = Auxiliary,
 }
-
-export abstract class StyleSheet {
-  protected readonly el: HTMLStyleElement | HTMLLinkElement
-
-  install(el: HTMLElement | ShadowRoot): void {
-    el.append(this.el)
-  }
-
-  uninstall(): void {
-    this.el.remove()
-  }
-
-  abstract to_native(): CSSStyleSheet
-}
-
-export class InlineStyleSheet extends StyleSheet {
-  protected override readonly el = style()
-
-  constructor(css?: string | CSSStyleSheetDecl, id?: string, readonly persistent: boolean = false) {
-    super()
-    if (isString(css)) {
-      this._update(css)
-    } else if (css != null) {
-      this._update(compose_stylesheet(css))
-    }
-    if (id != null) {
-      this.el.dataset.css = id
-    }
-  }
-
-  get css(): string {
-    return this.el.textContent
-  }
-
-  protected _update(css: string): void {
-    this.el.textContent = css
-  }
-
-  clear(): void {
-    this.replace("")
-  }
-
-  private _to_css(css: string, styles: CSSStyles | undefined): string {
-    if (styles == null) {
-      return css
-    } else {
-      return compose_stylesheet({[css]: styles})
-    }
-  }
-
-  replace(css: string, styles?: CSSStyles): void {
-    this._update(this._to_css(css, styles))
-  }
-
-  prepend(css: string, styles?: CSSStyles): void {
-    this._update(`${this._to_css(css, styles)}\n${this.css}`)
-  }
-
-  append(css: string, styles?: CSSStyles): void {
-    this._update(`${this.css}\n${this._to_css(css, styles)}`)
-  }
-
-  remove(): void {
-    this.el.remove()
-  }
-
-  to_native(): CSSStyleSheet {
-    const sheet = new CSSStyleSheet()
-    sheet.replaceSync(this.css)
-    return sheet
-  }
-}
-
-export class GlobalInlineStyleSheet extends InlineStyleSheet {
-  override install(): void {
-    if (!this.el.isConnected) {
-      document.head.appendChild(this.el)
-    }
-  }
-}
-
-export class ImportedStyleSheet extends StyleSheet {
-  protected override readonly el: HTMLLinkElement
-
-  constructor(url: string) {
-    super()
-    this.el = link({rel: "stylesheet", href: url})
-  }
-
-  replace(url: string): void {
-    this.el.href = url
-  }
-
-  remove(): void {
-    this.el.remove()
-  }
-
-  to_native(): CSSStyleSheet {
-    const sheet = new CSSStyleSheet()
-    sheet.replaceSync(`@import "${this.el.href}"`)
-    return sheet
-  }
-}
-
-export class GlobalImportedStyleSheet extends ImportedStyleSheet {
-  override install(): void {
-    if (!this.el.isConnected) {
-      document.head.appendChild(this.el)
-    }
-  }
-}
-
-export type StyleSheetLike = StyleSheet | string
 
 export async function dom_ready(): Promise<void> {
   if (document.readyState == "loading") {
