+++ bokehjs/src/lib/models/widgets/checkbox.tsx
@@ -1,11 +1,10 @@
 import {ToggleInput, ToggleInputView} from "./toggle_input"
-import type {StyleSheetLike} from "core/dom"
+import type {StyleSheetLike} from "core/stylesheets"
 import {ShadowComponent, cls} from "core/vdom"
 import type * as p from "core/properties"
 import checkbox_css from "styles/widgets/checkbox.css"
 import * as toggle_css from "styles/widgets/toggle_input.css"
 
-import {render} from "preact"
 import type {VNode} from "preact"
 import {signal} from "@preact/signals"
 import type {Signal} from "@preact/signals"
@@ -63,9 +62,9 @@ export class CheckboxView extends ToggleInputView {
     }
   }
 
-  component(): VNode {
+  override component(): VNode {
     const classes = [...this._css_classes()]
-    const stylesheets = this.adopted_stylesheets
+    const stylesheets = this.resolved_stylesheets
 
     const {active, label, disabled} = this.signals
     return (
@@ -75,12 +74,6 @@ export class CheckboxView extends ToggleInputView {
       </ShadowComponent>
     )
   }
-
-  override render(): void {
-    render(this.component(), this.el.parentNode!, this.el) // TODO preact-root-fragment
-  }
-
-  override readonly is_vdom = true
 }
 
 export namespace Checkbox {
